+++ UltraStar Play/Assets/Scenes/SongEditor/NoteArea/Drag/ManipulateNotesDragListener.cs
@@ -59,7 +59,7 @@ public class ManipulateNotesDragListener : MonoBehaviour, INeedInjection, IDragL
 
     public void OnBeginDrag(NoteAreaDragEvent dragEvent)
     {
-        if (dragEvent.GeneralDragEvent.InputButton != PointerEventData.InputButton.Left)
+        if (dragEvent.GeneralDragEvent.InputButton != (int)PointerEventData.InputButton.Left)
         {
             CancelDrag();
             return;
